style: merged AdminViewHeader and StudentViewHeader
